+++ server/src-lib/Hasura/RQL/DML/Delete.hs
@@ -95,6 +95,7 @@ runDelete
   :: ( HasVersion, QErrM m, UserInfoM m, CacheRM m
      , HasServerConfigCtx m, MonadIO m
      , Tracing.MonadTrace m, MonadBaseControl IO m
+     , MetadataM m
      )
   => Env.Environment
   -> DeleteQuery