+++ server/src-lib/Hasura/RQL/DML/Select.hs
@@ -334,7 +334,7 @@ phaseTwo =
 runSelect
   :: ( QErrM m, UserInfoM m, CacheRM m
      , HasServerConfigCtx m, MonadIO m, MonadBaseControl IO m
-     , Tracing.MonadTrace m
+     , Tracing.MonadTrace m, MetadataM m
      )
   => SelectQuery -> m EncJSON
 runSelect q = do